Mundanças cores e ajustes

+++ loja_app/models.py
@@ -33,7 +33,7 @@ class Produto(models.Model):
     """
     nome = models.CharField(max_length=100)
     descricao = models.TextField()
-    preco = models.IntegerField()
+    preco = models.DecimalField(max_digits=10, decimal_places=2)
     imagem = models.ImageField()
     categoria = models.ForeignKey(Categoria, on_delete=models.CASCADE)
